validasi kategori

+++ sipadi-Project/application/controllers/kategori/functions-kategori.php
@@ -94,6 +94,18 @@ function ubahKtg($data)
 function hapusKtg($id)
 {
     global $koneksi;
+
+    $qr_file = mysqli_query($koneksi, "SELECT gmbr FROM kategori WHERE id_kategori='$id'");
+    $hsl =  mysqli_fetch_array($qr_file);
+
+    if (!unlink("../../views/kategori/gambar/" . $hsl["gmbr"])) {
+        echo "<script>alert('error hapus gmbr');</script>";
+        return false;
+    }
+
+
+
+
     mysqli_query($koneksi, "DELETE FROM kategori WHERE id_kategori='$id'");
 
     return mysqli_affected_rows($koneksi);
